compositor|internal: Reach schema modules directly from the entry
A barrel belongs only at a package's published entry point, and the exports map names `.` and `./samples/*.json` alone. `src/schemas/index.ts` was neither, and `src/index.ts` reached its schemas through it, so importing `composeArtifactId` loaded every schema module. The entry now re-exports from the defining modules, and the intermediate barrel is gone.

The prose that barrel carried in its section headers moves onto the modules it described, where a reader opening one finds it.

`samples/index.ts` becomes `samples/sample-documents.ts`. It is not a re-export module: it holds `buildSampleDocuments` and `renderSampleJson`, which the generator script and two test files call. Because nmr's coverage config excludes `**/index.ts` as pure re-exports, both functions went unmeasured while their two siblings in the same directory were covered. The dead re-export tail goes with the rename -- nothing reached the sample builders through it.

The barrel's completeness check moves up a level with the surface it guards, so a schema added without an entry export still fails.
