+++ packages/compositor/src/schemas/__tests__/plan-schema.unit.test.ts
@@ -1,26 +1,10 @@
 import { describe, expect, it } from 'vitest';
 
 import { buildPlan } from '../../test-utils/buildPlan.ts';
-import * as commonSchemas from '../common.ts';
-import * as descriptorSchemas from '../descriptor-schemas.ts';
-import * as fileSchemas from '../file-schemas.ts';
-import * as graphSchemas from '../graph-schemas.ts';
-import * as barrel from '../index.ts';
-import * as planSchema from '../plan-schema.ts';
 import { PLAN_SCHEMA_VERSION, PlanSchema } from '../plan-schema.ts';
-import * as resolutionSchemas from '../resolution-schemas.ts';
 import { findIssuePaths } from '../test-utils/findIssuePaths.ts';
 
 const tables = ['artifacts', 'blobs', 'files', 'fingerprint', 'kinds', 'partials', 'sources', 'targets'] as const;
-
-const modules: ReadonlyArray<readonly [string, Record<string, unknown>]> = [
-  ['common', commonSchemas],
-  ['descriptorSchemas', descriptorSchemas],
-  ['fileSchemas', fileSchemas],
-  ['graphSchemas', graphSchemas],
-  ['planSchema', planSchema],
-  ['resolutionSchemas', resolutionSchemas],
-];
 
 describe('PlanSchema', () => {
   it('accepts a plan carrying every table', () => {
@@ -47,14 +31,5 @@ describe('PlanSchema', () => {
 
   it('declares a positive schema version', () => {
     expect(PLAN_SCHEMA_VERSION).toBeGreaterThan(0);
-  });
-});
-
-describe('schema barrel', () => {
-  it.each(modules)('re-exports every runtime member of %s', (_label, module) => {
-    const exported = new Set(Object.keys(barrel));
-    const missing = Object.keys(module).filter((name) => !exported.has(name));
-
-    expect(missing).toStrictEqual([]);
   });
 });
